Use kubelet credentials and rely on NodeRestriction
By relying on NodeRestriction Admission controller for security,
we can make sure the daemonset instance is not touching something
it should not. This improves the security, at the expense of
exposing the kubelet configuration and credentials to kured.

The default location of the kubelet is the one provided by
kubeadm. For non-kubeadm or custom kubeadm clusters, a user can
provide the kubelet configuration file path, which should contain
all the information necessary for kubelet to issue commands on the
kubernetes api servers.