Add notifyURL support in helm chart
Without this patch, notifyURL will not be taken into consideration
when deploying with the helm chart, because the value is not
templated and not exposed to end users.

This is a problem because the end users only have access to the
previous kind of notifications (slack arguments).

This fixes it by modifying the template, and exposing a default
empty value for notifyURL.

Co-Authored-By: Alexei Tighineanu <atighineanu@suse.de>