Simplify lock state
Implementation details of lock should not leak into the calling
methods. Without this, calls are a bit more complex
 and error handling is harder to find.

I solve this by inlining most of the error handling in the
function where we bubble up all errors and including all the
methods for managing locks into the daemonsetlock package.

This will help us add in new kind of locks in the future,
without much impact on main. The daemonset packages will
be simpler this way. By that point, we should probably rename
the package to lock, instead of daemonsetlock...

Signed-off-by: Jean-Philippe Evrard <open-source@a.spamming.party>