Make locks more generic
Implementation details of lock should not leak into the calling
methods, and lock purpose or authoritative data is not always
clear.

Without this path, calls are a bit more complex
and error handling is harder to find.

I solve this by inlining most of the error handling in the
function where we bubble up all errors and including all the
methods for managing locks into the daemonsetlock package.

This will help us add in new kind of locks in the future,
without much impact on main. The daemonset packages will
be simpler this way. By that point, we should probably rename
the package to lock, instead of daemonsetlock...

This allowed me to find that the holding method not only
used to test that a lock was held, but also extracting
the information required in terms of node metadata
(which here only means unschedulable).

As I am not too fond of silently mutating data,
I picked to explicitly return the lock data instead.

I also made it explicit that the Acquire lock method
is passing the node metadata as structured information,
rather than an interface{}. This is a bit more fragile
at runtime, but I prefer having very explicit errors if
the locks are incorrect, rather than having to deal with
unvalidated data.

Signed-off-by: Jean-Philippe Evrard <open-source@a.spamming.party>