Cleanup the main loop
This patches simplifies the principal loop, and integrate
the prometheus metrics into the main loop.

Without this, some person might see the /metrics endpoint
telling a reboot is required, but the principal go routine not
even there yet, and not even checking the reboot.

This is a problem, as it has lead to misunderstanding in the past,
and we even clarified some of them into our main documentation.

Yet, some ppl still miss that part of the documentation, and find
the period a bit of a misnomer.

This fixes it by merging the maintainRebootRequiredMetric
with rebootAsRequired method.

However, this also means a change in the defaults, to align
the expectations. The period had to be reduced in order to
match the prometheus expectations. This means that we will
check more actively for reboot information into
rebootAsRequired method. I consider this is fine, as it is
better to tick more frequently to know what to do.

This might have an impact on large clusters, which will do more
API calls.

In order to fix that, I reschuffled the main loop, to
focus on whether or not a reboot is required at each tick,
which helps a bit reducing some of the calls.

If this is still a problem, we could find a way to pass
node and daemonset information in the different calls inside
a tick, to make it a bit more efficient.

A new struct GenericRebooter was introduced to keep data
all the rebooters need. For example, here, it keeps the
rebootDelay imposed to the reboot function. Each rebooter
can implement its own reboot method, but to simplify there
the generic rebooter implements a DelayReboot() method,
which is then called from all the Reboot methods of the
concrete types.

On the way, this commit also changed the RebootBlocked function,
to also return the list of the names of (types) blocking the
reboot.

This is useful to debug, but also can be useful to later
expose the content of the blocker into specific metrics.

Signed-off-by: Jean-Philippe Evrard <open-source@a.spamming.party>