Make lock release delay part of lock info
Without this, the lock release delay is part of the
rebootasrequired loop, where it makes sense to be part of the
lock itself.

Instead of passing the argument as parameter of the
rebootasrequired method, this refactor took creation of the lock
object in the main loop, close to all the variables, and then
pass the lock object to the rebootasrequired. This makes the
call for rebootasrequired more clear, and lock is now
encompassing everything needed to acquire, release, or get
info about the lock.