Modify lock interfaces
This commit modifies the acquire and holding
interfaces, as they were not behaving in an idiomatic way:

1. Acquire needed a test of holding beforehand.

This is a problem, as acquire should simply try to acquire
the lock or fail. By implementing this, we simplify our
rebootasrequired loop.

2. Acquire method returned who was holding the lock,
depending on whether acquiry was successful or not.

This is not really a problem, but it extends the Acquire
method to something it was not really expected to do,
and for which the results were unexpected for anyone
not reading the code of daemonsetlock.

I fixed it by simply removing the data of who is holding
the lock. It does not make any value whatsoever to list
in the log who is holding the lock, simply knowing that
it's blocked because of the absence of aquiry of lock
is enough. Someone wanting to know can still debug
by fetching the node information live.

3. Holding method is poorly named compared to what the
method actually does: It does not only return the holding
information, but also who is holding the lock.

I attempted to resolve this by spliting the method in
two: One actually giving Holding information, and one
getting the rest of the data. It resulted in lots
of duplicated code, and mostly the validity of the
results fetched needed an extra field: two fields
are necessary: Which node does the metadata applies,
and the medata itself.

Hence the method is useful for both cases, but the
naming is a bit off. I renamed the method into
GetNodeData() as it clarifies the intent:
- If holding the lock, you get the node data;
- If not holding the lock, you get empty information;
- If an happened on the way, you know you can retry.

4. Release() is not idempotent

Without this Release() would have errors if there was nothing
to release, or there was no lock.

If Release() means removing the lock regardless of previous state,
the code can be cleaned up further.

Signed-off-by: Jean-Philippe Evrard <open-source@a.spamming.party>