Remove util package
This is not easy to browse and does not make any purpose.

This makes it marginally better for privileged call, which
at the end could as well be in each package for small repetition.

If really necessary, we could rebuild this as an internal package,
but I don't see why we should bother.

Signed-off-by: Jean-Philippe Evrard <open-source@a.spamming.party>