Fix discrepency metrics <> rebootAsRequired
Without this patch, one metric could say "reboot is required"
while the rebootAsRequired tick did not run (long period for
example).

This is a problem, as it leads to misexpectations: "Why
did the system not reboot, while the metrics indicate a reboot
was required".

This solves it by inlining the metrics management within the
rebootAsRequired goroutine.

Closes: #725

Signed-off-by: Jean-Philippe Evrard <open-source@a.spamming.party>