Expose unpriv commands
Wthout this, it will not be possible to use unprivileged
commands.

This is a problem, as one might want to run a command
without the nsenter to pid 1.

This fixes it by exposing this to main, the only thing
remaining is to use a boolean to expose the feature in
a later commit.

Signed-off-by: Jean-Philippe Evrard <open-source@a.spamming.party>