Move to slog
Without this, we are using logrus and the logs are inconsistent.

This fixes it by moving to slog, and reviewing all the logs.

I added multiple comments to clarify the intent behind the logs.

With structured logging, extra data, such as the node, is exposed
whenever possible, which makes the logging more consistent.

We are not yet using contextes for ticks, it should be done in
a later commit.

Signed-off-by: Jean-Philippe Evrard <open-source@a.spamming.party>