Fix typo in signal name
signal is named SIGRTMIN not SIGTRMIN. It can be confusing
for people.

As this could also lead to questions, the code clarifies a
TODO item: We are still relying on hardcoded ints, instead
of evaluating the signal number at run time, as recommended
by man (7) signal. Please read man (7) signal for further
details.

Signed-off-by: Jean-Philippe Evrard <open-source@a.spamming.party>