Extract flags which can be reused
Without this, we will need to repeat the pflag handling in each
cmd.

This is a problem, as it will lead to discrepencies into the
handling of env vars or cli flags.

This fixes it by moving to an internal package, allowing them
to be reused across the different CLIs.

Signed-off-by: Jean-Philippe Evrard <open-source@a.spamming.party>